requirement.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
statsmodels>=0.14
scipy>=1.10
requests>=2.28
scikit-learn>=1.3
python-dotenv>=1.0     
ruptures>=1.1
pmdarima>=2.0 
